+++ pajbot/modules/actionchecker.py
@@ -79,13 +79,23 @@ class ActionCheckerModule(BaseModule):
             default=500,
             constraints={"min_value": 100, "max_value": 1000},
         ),
+        ModuleSetting(
+            key="disable_warnings",
+            label="Disable warning timeouts",
+            type="boolean",
+            required=True,
+            default=False,
+        ),
     ]
 
     def delete_or_timeout(self, user, msg_id, reason):
         if self.settings["moderation_action"] == "Delete":
             self.bot.delete_message(msg_id)
         elif self.settings["moderation_action"] == "Timeout":
-            self.bot.timeout(user, self.settings["timeout_length"], reason, once=True)
+            if self.settings["disable_warnings"] is True:
+                self.bot.timeout(user, self.settings["timeout_length"], reason, once=True)
+            else:
+                self.bot.timeout_warn(user, self.settings["timeout_length"], reason, once=True)
 
     def on_message(self, source, message, event, msg_id, **rest):
         if self.settings["enabled_by_stream_status"] == "Online Only" and not self.bot.is_online: